==> code/1_collection.ipynb <==
## Using the Spoonacular API through rapidapi to gather random recipe data 

import requests
import json
import os

api_key = "75791d1213msh141b4116de88e2fp11abe7jsn18de53e5244b"

# function to call each day
# spoonacular provides 500 points free per day callable 100 at a time
# times = number of times to make the call
# num = points to call each time

def spoonacular_random(times=4, num=100):
    url = "https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/random"
    querystring = {"number":"100"}
    headers = {
        'x-rapidapi-key': api_key,
        'x-rapidapi-host': "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com"
        }
    start = int(os.listdir('../data/api_data/')[-1][0:3]) + 1
    
    for c in range(times): 
        response = requests.request("GET", url, headers=headers, params=querystring)
        data = response.json()
        current_iter = str(start + c).zfill(3)
        with open(f'../data/api_data/{current_iter}.json', 'w') as f:
            json.dump(data, f)

spoonacular_random()

---
### Manual calling
Commented out so running the entire notebook doesn't incur charges

# # using pregenerated api call posted on rapidapi.com
# url = "https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/random"

# querystring = {"number":"100"}

# headers = {
#     'x-rapidapi-key': api_key,
#     'x-rapidapi-host': "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com"
#     }

# # the call
# response = requests.request("GET", url, headers=headers, params=querystring)
# # the response
# response
# data = response.json()

# # save each
# current_iter = '13'

# with open(f'../data/api_data/{current_iter}.json', 'w') as f:
#     json.dump(data, f)
