==> code/2_cleaning.ipynb <==
## Cleaning

import json
import os
import pandas as pd
import numpy as np
import re

from sklearn.feature_extraction.text import CountVectorizer

---
### Load in all the recipes

# verifies the folder
os.listdir('../data/api_data/')

# load in all json as dataframes
list_of_dfs = []
for file in os.listdir('../data/api_data/')[1:]:
    with open(f'../data/api_data/{file}') as f:
        list_of_dfs.append(pd.DataFrame(json.load(f)['recipes']))

---
### Merge all the dataframes into one, dropping duplicates

master_df = list_of_dfs[0]

for dframe in list_of_dfs[1:]:
    master_df = master_df.append(dframe)

# random recipes were acquired using the api
# drop duplicates
master_df = master_df.drop_duplicates(subset=['id'])

master_df.shape

---
### Clean up json format columns

json_columns = [
    'extendedIngredients',
    'analyzedInstructions'
]

Simplified list of ingredients

master_df['simplifiedIngredients'] = [[ing for ing in pd.DataFrame(cell)['nameClean']] for cell in master_df['extendedIngredients']]

Simplified list of instructions

# needed a function to pull all the steps from each component of the recipe
def make_list(df_row):
    list_of_steps = []
    for a in range(len(df_row)):
        for b in range(len(df_row[a]['steps'])):
            list_of_steps.append(df_row[a]['steps'][b]['step'])
    return list_of_steps

master_df['simplifiedInstructions'] = [make_list(row) for row in master_df['analyzedInstructions']]

# # example of needing to go multiple layers down, get the steps, then switch higher up to a separate component of the meal to get its steps
# for a in range(len(master_df['analyzedInstructions'].iloc[0])):
#     for b in range(len(master_df['analyzedInstructions'].iloc[0][a]['steps'])):
#         print(master_df['analyzedInstructions'].iloc[0][a]['steps'][b]['step'])

---
### A bit of feature selection to make the dataframe a bit smaller

master_df.columns

# While some models may drop further features, these are the ones I don't plan on using at all
features_to_drop = [
    'weightWatcherSmartPoints',
    'gaps',
    'veryPopular',
    'aggregateLikes',
    'spoonacularScore',
    'healthScore',
    'creditsText',
    'license',
    'sourceName',
    'pricePerServing',
    'id',
    'servings',
    'sourceUrl',
    'image',
    'imageType',
    'diets',
    'originalId',
    'spoonacularSourceUrl',
    'cookingMinutes',
    'preparationMinutes',
    'author',
    'extendedIngredients',
    'analyzedInstructions'
]

# drop the features
master_df = master_df.drop(columns=features_to_drop)

# dropping the columns had the added benefit of clearing all of the missing values
# one thing to note however is that some of the columns have an empty list instead of nan
# those were not removed
master_df.isnull().sum().sort_values(ascending=False)[:3]

---
### Convert booleans to numerics

boolean_columns = [
    'vegetarian',
    'vegan',
    'glutenFree',
    'dairyFree',
    'veryHealthy',
    'cheap',
    'sustainable',
    'lowFodmap'
]

for column in boolean_columns:
    master_df[column] = [1 if val is True else 0 for val in master_df[column]]

---
## TODO: Dummify dishType
so it doesn't break when exported as a csv and reimported elsewhere

master_df['dishTypes'].iloc[400][0]

# example of the original
print(type(master_df['dishTypes'].iloc[400]))
master_df['dishTypes'].iloc[400]

---
## Simplify the cuisines column

simplifiedCuisine = {
    '[]': '',
    "['Mexican']": 'Mexican',
    "['Mediterranean', 'Italian', 'European']": 'Mediterranean',
    "['American']": 'American',
    "['German', 'European']": 'German',
    "['Mediterranean', 'French', 'European']": 'French',
    "['Creole', 'Cajun']": 'Cajun',
    "['Southern']": 'American',
    "['Chinese', 'Asian']": 'Chinese',
    "['Asian']": 'Asian',
    "['English', 'Scottish', 'British', 'European']": 'British',
    "['Mediterranean', 'European', 'Greek']": '',
    "['Indian', 'Asian']": 'Indian',
    "['African']": 'African',
    "['Korean', 'Asian']": 'Korean',
    "['Vietnamese', 'Asian']": 'Vietnamese',
    "['Spanish', 'European']": 'Spanish',
    "['Middle Eastern']": 'Middle Eastern',
    "['Eastern European', 'European']": 'Eastern European',
    "['European', 'Irish']": 'Irish',
    "['Cajun', 'Creole']": 'Cajun',
    "['South American', 'Latin American']": 'South American',
    "['Thai', 'Asian']": 'Thai',
    "['Jewish']": 'Middle Eastern',
    "['Japanese', 'Asian']": 'Japanese'
}

master_df['cuisines'] = [str(val) for val in master_df['cuisines']]
master_df['cuisines'] = master_df['cuisines'].map(simplifiedCuisine)

---
## Convert the ingredients from list to string form

master_df['simplifiedIngredients']

[str]

---
### Save as a csv

master_df.to_csv('../data/recipes.csv')
