==> code/4.1_kmeans.ipynb <==
## KMeans classification

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer

from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score

df = pd.read_csv('../data/recipes.csv').drop(columns=['Unnamed: 0'])
print(df.shape)
df['count'] = 1
df.head(3)

to_drop = [
    'cuisines', 
    'occasions',
    'simplifiedIngredients',
    'simplifiedInstructions'
]

df = df.drop(columns=to_drop)

---
### Model using summary

sum_df = df['summary']

sum_df

cv = CountVectorizer(stop_words='english')
cv.fit(sum_df)
X = cv.transform(sum_df)

len(cv.get_feature_names())

'chicken' in cv.get_feature_names()

pd.DataFrame.sparse.from_spmatrix(X, columns=cv.get_feature_names())['chicken'].sort_values(ascending=False).head(135)

km = KMeans(n_clusters=4, random_state=42)
km.fit(X)

preds = km.predict(X)

df['pred_by_summary'] = preds

df.groupby(by='pred_by_summary').sum()

---
### Model using instructions

ins_df = df['instructions']

ins_df = ins_df.fillna('')

ins_df

cv = CountVectorizer(stop_words='english')
cv.fit(ins_df)
X = cv.transform(ins_df)

X

len(cv.get_feature_names())

km = KMeans(n_clusters=4, random_state=42)
km.fit(X)

preds = km.predict(X)

df['pred_by_instructions'] = preds

df.groupby(by='pred_by_instructions').sum()

---
### Model using summary and instructions

com_df = df['summary'] + df['instructions'].fillna('')

com_df

type(com_df.iloc[0])

cv = CountVectorizer()
cv.fit(com_df)
X = cv.transform(com_df)

len(cv.get_feature_names())

# Check how many occurances of chicken there are in the recipe ingredients
pd.DataFrame.sparse.from_spmatrix(X, columns=cv.get_feature_names())['chicken'].sort_values(ascending=False).head(160)

# someone really likes chicken
df.iloc[87]

df['simplifiedIngredients'].iloc[87]

km = KMeans(n_clusters=4, random_state=42)
km.fit(X)

preds = km.predict(X)

df['pred_by_summary'] = preds

df.groupby(by='pred_by_summary').sum()
