==> code/4.3_markov.ipynb <==
## Markov Chain instruction generation

### Imports

import pandas as pd
import re
import markovify
import nltk
import spacy

# !python -m spacy download en

df = pd.read_csv('../data/recipes.csv').drop(columns=['Unnamed: 0'])
print(df.shape)

---
### Cleaning for markovify

# text cleaning function credit: https://towardsdatascience.com/creating-word-embeddings-coding-the-word2vec-algorithm-in-python-using-deep-learning-b337d0ba17a8
def clean_text(
    string: str, 
    punctuations=r'''!()-[]{};:'"\,<>/?@#$%^&*_~''',
    stop_words=['the', 'a', 'and', 'is', 'be', 'will']) -> str:
    """
    A method to clean text 
    """
    # Cleaning the urls
    string = re.sub(r'https?://\S+|www\.\S+', '', string)

    # Cleaning the html elements
    string = re.sub(r'<.*?>', '', string)

    # Removing the punctuations
    for x in string.lower(): 
        if x in punctuations: 
            string = string.replace(x, "") 

    # Converting the text to lower
    string = string.lower()

    # not this time
    # Removing stop words
    # string = ' '.join([word for word in string.split() if word not in stop_words])

    # Cleaning the whitespaces
    string = re.sub(r'\s+', ' ', string).strip()
    
    # Remove linebreak
    string = string.replace('\r\n', '')

    return string  

df.head(1)

df['instructions'] = [clean_text(str(row)) for row in df['instructions']]

df['instructions']

---
### Parse

Originally going to use [spaCy](https://spacy.io/usage/models), but did it manually above in cleaning instead

# accuracy
# nlp = spacy.load("en_core_web_sm")

# efficiency
nlp = spacy.load('en_core_web_sm')

recipe_soup = ''
for instr in df['instructions']:
    recipe_soup += instr
# verify: oh it worked
# recipe_soup

---
### Markovify model

# 'fit' the markovify model
mark = markovify.Text(recipe_soup, state_size=2)

print(mark.make_sentence()[0:500])
